Muchas fuciones agregadas. Pronto un changelog
