+++ .client_secrets.encrypt
@@ -1,0 +1,1 @@
+{"web": {"client_id": "12218029133-o1tvim2bi27v1tsrrpkh210603nc70nn.apps.googleusercontent.com", "project_id": "pyfilefinder", "auth_uri": "https://accounts.google.com/o/oauth2/auth", "token_uri": "https://oauth2.googleapis.com/token", "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs", "client_secret": "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", "redirect_uris": ["http://localhost:8080/"], "javascript_origins": ["http://localhost:8080"]}}
