+++ modules/funcs.py
@@ -7,9 +7,15 @@ from pydrive2.drive import GoogleDrive
 from statistics import median
 from datetime import datetime
 from shutil import move as mv
+from google.oauth2.credentials import Credentials
+from googleapiclient.discovery import build
 
 
 def compress_path(path: list):
+    """
+        Funcion incompleta
+        path: Objeto con un directorio adentro
+    """
     media = int(median(range(len(path))))
     for x,v in enumerate(path):
         if x == 0:
@@ -28,19 +34,40 @@ def compress_path(path: list):
     if len(path[len(path) - 1]) > 35:
         path[len(path) - 1] = path[len(path) - 1][0:35] + "..."
     return path
-    
+
+compress_path()
 
 class jsonEx:
+    """
+        Clase para dar soporte a los archivos de configuracion y la base de datos que contiene las extensiones
+    """
     def update_json(filex,data) -> None:
+        """
+            Actualiza los datos de un archivo json. No devuelve nada
+            filex: Archivo json
+            data: Diccionario con configuraciones adentro
+        """
         with open(filex,'w') as file:
             json.dump(data, file, indent=4)
         
-    def get(filex):
+    def get(filex) -> dict:
+        """
+            Obtiene los datos de un archivo json. Devuelve dict
+            filex: Archivo json
+        """
         with open(filex,'r') as file:
             return json.load(file)
 
 class termui:
-    def terminal(opts, return_type=bool):
+    """
+        Clase para facilitar la escritura de un TUI
+    """
+    def terminal(opts, return_type=bool) -> int | bool:
+        """
+            Muestra un menu interactivo. Devuelve int o bool, por predeterminado int
+            opts: Una lista conteniendo las opciones
+            return_type (valor predeterminado bool): Especifica que quieres que se devuelva, bool para indicar si el primer elemento de la lista fue seleccionado e int para indicar que elemento de la lista se selecciono
+        """
         if len(opts) < 1:
             raise IndexError("debug")
         menu = TerminalMenu(opts)
@@ -50,8 +77,13 @@ class termui:
             return menu.show()
         else:
             raise TypeError('Incompatible type')
+    
+termui().terminal()
 
 class compress():
+    """
+        Clase para la facilitacion de la compresion de archivos, pronto con soporte para otros formatos
+    """
     def __init__(self, filename : str, format="zip"):
         filename = os.path.splitext(filename)[0]
         if format == "zip":
@@ -71,6 +103,11 @@ class compress():
                 print(f"{label} ({current}/{max}) {int(current/max * 100)}%", end='\r')
 
     def add(self, files : list, callback=_progress):
+        """
+            Agrega archivos a un fichero zip. No devuelve nada
+            files: Una lista conteniendo los archivos a comprimir
+            callback=_progress: Un callback que se utilizara para mostrar el progreso, por defecto hay una funcion simple que muestra el progreso
+        """
         if callback == None:
             for v,x in enumerate(files):
                 self.file.write(x)
@@ -81,12 +118,22 @@ class compress():
         print("")
     
     def add_single(self, files: str):
+        """
+            Agrega un solo archivo a un fichero zip. No devuelve nada
+            files: Un archivo para ser comprimido
+        """
         self.file.write(files)
     
     def finish(self):
+        """
+            IMPORTANTE EJECUTAR ESTO ya que si no lo ejecutas, el fichero estara incompleto
+        """
         self.file.close()
 
 class finder():
+    """
+        Clase para buscar archivos en el disco
+    """
     def __init__(self, path, ext_list) -> None:
         if os.path.basename(path) != '':
             path = path + '/'
@@ -95,6 +142,9 @@ class finder():
         self.exts = jsonxd[ext_list]
     
     def find(self) -> dict:
+        """
+            Devuelve un diccionario conteniendo las extensiones junto con una lista de los archivos encontrados con su extension correspondiente
+        """
         founded = {}
         self.all = []
         for x in self.exts:
@@ -109,6 +159,10 @@ class finder():
         return founded
     
     def find_userext(self,userext) -> dict:
+        """
+            Devuelve un diccionario conteniendo los archivos encontrados con la extension especificada
+            userext: Extension con la que quieras buscar los archivos
+        """
         founded = {}
         self.all = []
         path = glob.glob(f"{self.path}**/*.{userext}", recursive=True)
@@ -119,6 +173,11 @@ class finder():
         return founded
     
     def find_in_list(self,word,strict=False) -> list:
+        """
+            La funcion busca entre los archivos ya encontrados con la funcion find() con la variable word. Devuelve una lista
+            word: Palabra por filtrar entre los archivos
+            strict=False: Por defecto es falso, si es verdadero se usara una funcion para encontrar mas estrictamente entre los archivos
+        """
         if not strict:
             word = os.path.splitext(word)[0]
             result = filter(lambda x: re.search("{}.*".format(word), x), self.all)
@@ -132,6 +191,9 @@ class finder():
             return real
     
     def drive_format(self) -> dict:
+        """
+            Funcion para darle un formato especifico para usar con la clase Drive ({'names':[], 'paths':[]}). Devuelve dict
+        """
         dictionary = {
             'names':[],
             'paths':[]
@@ -142,6 +204,9 @@ class finder():
         return dictionary
 
     def generalize(self) -> list:
+        """
+            Funcion para convertir el diccionario con los archivos ordenados por extensiones a una lista. Devuelve una lista
+        """
         return self.all
 
     def __findstrict(self,word) -> list:
@@ -154,15 +219,31 @@ class finder():
         return ret
 
 class files_handler:
+    """
+        Clase de funciones varias para manejar archivos
+    """
     def wipe_txt(file: str):
+        """
+            Limpia un archivo de texto. No devuelve nada
+            file: Ruta del archivo de texto
+        """
         open(file,'w').close()
         
     def stat(file: str) -> list:
+        """
+            Funcion para devolver metadatos de un archivo. Devuelve list
+            file: Archivo a analizar
+        """
         stated = os.stat(file)
         name = os.path.basename(file)
         ext = os.path.splitext(name)
         return [name,file,ext[1],datetime.fromtimestamp(stated.st_mtime).strftime('%Y-%m-%d %H:%M'), stated.st_size / 1000000, stated.st_uid]
     def save(file: str, content: str | list):
+        """
+            Guarda un archivo con el contenido que especifiques (El contenido anterior del archivo se elimina). No devuelve nada
+            file: Archivo a escribir
+            content: Contenido 
+        """
         if type(content) == list:
             files_handler.wipe_txt(file)
             with open(file,'a') as file:
@@ -182,6 +263,9 @@ class files_handler:
 
 
 class Drive():
+    """
+        Clase para automatizar el logeo de Google Drive
+    """
     def __init__(self, file=os.path.join(os.getenv('HOME'), '.local/share/secrets/.pydrive/.credentials.encrypt')):
         if os.path.exists(file):
             self.gauth = GoogleAuth()
@@ -191,16 +275,19 @@ class Drive():
             self.gauth.LoadClientConfigFile(os.path.join(os.getenv('HOME'), '.local/share/secrets/.pydrive/.client_secrets.encrypt'))
             self.drive = GoogleDrive(self.gauth)
             os.system("./tests/main --create")
+            self.credentials = file
         else:
             os.system("./tests/main --create")
             os.system("./tests/main --load")
             self.gauth = GoogleAuth()
             self.gauth.LoadClientConfigFile(os.path.join(os.getenv('HOME'), '.local/share/secrets/.pydrive/.client_secrets.encrypt'))
             self.gauth.LocalWebserverAuth()
-            os.system("./tests/main --create")
+            #os.system("./tests/main --create")
+            print(file)
             os.system(f'mkdir -p {os.path.split(file)[0]}')
             self.gauth.SaveCredentialsFile(file)
             self.drive = GoogleDrive(self.gauth)
+            self.credentials = file
         
     def __callbacks(current,max,label=""):
         if current == max - 1:
@@ -208,8 +295,17 @@ class Drive():
         else:
             print(f"{label} ({current}/{max}) %{int(current/max * 100)}", end='\r')
     
+    def user_info(self) -> dict:
+        """
+            Obtiene un diccionario conteniendo informacion del usuario. Devuelve dict
+        """
+        return self.drive.GetAbout()
 
-    def create_folder(self,folder: str):
+    def create_folder(self,folder: str) -> int | str:
+        """
+            Crea una carpeta en la nube del usuario, si ya existe devuelve el id. Devuelve algo (No estoy seguro)
+            folder: Nombre de la carpeta a crear
+        """
         folder_list = self.drive.ListFile({'q': "mimeType='application/vnd.google-apps.folder' and trashed = false and title='"+folder+"'"}).GetList()
         if len(folder_list)>0:
             return folder_list[0]['id']
@@ -218,18 +314,33 @@ class Drive():
             folder_list.Upload()
             return folder_list['id']
 
-    def upload(self, filexd: dict, folder: str, label="Subiendo..."):
-        filex = self.drive.CreateFile({'title': filexd['names'], 'parents': [{'id': folder}]})
-        filex.SetContentFile(filexd['paths']) # especificar la ruta del archivo en tu computadora
-        print(label)
+    def upload(self, name: str, content: str, folder: str, label="Subiendo...", endx='\n'):
+        """
+            Funcion para subir un solo archivo a Google Drive. No devuelve nada
+            name: Nombre del archivo
+            content: Ruta del archivo
+            folder: ID de la carpeta en donde se subira el archivo
+            label="Subiendo...": Etiqueta que se imprimira en la terminal, puedes dejarlo como una cadena vacia
+            endx='\n': Se pondra al final de print(), puedes dejarlo como una cadena vacia
+        """
+        filex = self.drive.CreateFile({'title': name, 'parents': [{'id': folder}]})
+        filex.SetContentFile(content) # especificar la ruta del archivo en tu computadora
+        print(label, end=endx)
         filex.Upload()
     
     def uploads(self, files: dict, folder: str, callback=__callbacks):
+        """
+            Funcion para subir varios archivos a Google Drive. No devuelve nada
+            files: Diccionario conteniendo los nombres y las rutas de los archivos (Ejemplo: {'names':[], 'paths':[]})
+            folder: ID de la carpeta en donde se subiran los archivos
+            callback=__callbacks: Funcion callback que se utilizara para mostrar el progreso de subida de archivos
+        """
         for v,x in enumerate(files['names']):
             file = self.drive.CreateFile({'title': x, 'parents': [{'id': folder}]})
             file.SetContentFile(files['paths'][v]) # especificar la ruta del archivo en tu computadora
-            file.Upload()
             callback(v,len(files['names']),label="Subiendo archivos")
+
+            file.Upload()
 
 def printc(msg,color, endx='\n', label="[INFO]"):
     print(Colorate.Horizontal(color, label), msg, end=endx)
