+++ test.py
@@ -1,104 +1,10 @@
-import gi
+from modules.funcs import Drive
 
-gi.require_version("Gtk", "3.0")
-from gi.repository import Gtk
+info = Drive().user_info()
 
+for x in info:
+    print("Nombre: ",x, "Tipo: ",type(info[x]), "Salto de linea: \n")
+    print(info[x])
 
-class MessageDialogWindow(Gtk.Window):
-    def __init__(self):
-        super().__init__(title="MessageDialog Example")
-
-        box = Gtk.Box(spacing=6)
-        self.add(box)
-
-        button1 = Gtk.Button(label="Information")
-        button1.connect("clicked", self.on_info_clicked)
-        box.add(button1)
-
-        button2 = Gtk.Button(label="Error")
-        button2.connect("clicked", self.on_error_clicked)
-        box.add(button2)
-
-        button3 = Gtk.Button(label="Warning")
-        button3.connect("clicked", self.on_warn_clicked)
-        box.add(button3)
-
-        button4 = Gtk.Button(label="Question")
-        button4.connect("clicked", self.on_question_clicked)
-        box.add(button4)
-
-    def on_info_clicked(self, widget):
-        dialog = Gtk.MessageDialog(
-            transient_for=self,
-            flags=0,
-            message_type=Gtk.MessageType.INFO,
-            buttons=Gtk.ButtonsType.OK,
-            text="This is an INFO MessageDialog",
-        )
-        dialog.format_secondary_text(
-            "And this is the secondary text that explains things."
-        )
-        dialog.run()
-        print("INFO dialog closed")
-
-        dialog.destroy()
-
-    def on_error_clicked(self, widget):
-        dialog = Gtk.MessageDialog(
-            transient_for=self,
-            flags=0,
-            message_type=Gtk.MessageType.ERROR,
-            buttons=Gtk.ButtonsType.CANCEL,
-            text="This is an ERROR MessageDialog",
-        )
-        dialog.format_secondary_text(
-            "And this is the secondary text that explains things."
-        )
-        dialog.run()
-        print("ERROR dialog closed")
-
-        dialog.destroy()
-
-    def on_warn_clicked(self, widget):
-        dialog = Gtk.MessageDialog(
-            transient_for=self,
-            flags=0,
-            message_type=Gtk.MessageType.WARNING,
-            buttons=Gtk.ButtonsType.OK_CANCEL,
-            text="This is an WARNING MessageDialog",
-        )
-        dialog.format_secondary_text(
-            "And this is the secondary text that explains things."
-        )
-        response = dialog.run()
-        if response == Gtk.ResponseType.OK:
-            print("WARN dialog closed by clicking OK button")
-        elif response == Gtk.ResponseType.CANCEL:
-            print("WARN dialog closed by clicking CANCEL button")
-
-        dialog.destroy()
-
-    def on_question_clicked(self, widget):
-        dialog = Gtk.MessageDialog(
-            transient_for=self,
-            flags=0,
-            message_type=Gtk.MessageType.QUESTION,
-            buttons=Gtk.ButtonsType.YES_NO,
-            text="This is an QUESTION MessageDialog",
-        )
-        dialog.format_secondary_text(
-            "And this is the secondary text that explains things."
-        )
-        response = dialog.run()
-        if response == Gtk.ResponseType.YES:
-            print("QUESTION dialog closed by clicking YES button")
-        elif response == Gtk.ResponseType.NO:
-            print("QUESTION dialog closed by clicking NO button")
-
-        dialog.destroy()
-
-
-win = MessageDialogWindow()
-win.connect("destroy", Gtk.main_quit)
-win.show_all()
-Gtk.main()
+print()
+print("Informacion especifica:", info['user']['picture']['url'])
